T7737: add requests to provide auxiliary set/delete operations on config
Occasionally, it is useful to update the config tree _during a commit_
with non-actionable data (meaning data not required for the actions of
currently invoked config_mode scripts): the classic example is that of
removing the entries of plaintext passwords and adding the encrypted
entry. Add the relevent requests and session record field.